refactor(domain): move out logic for updating monitor state to new `RefreshMonitorStateCommand`
