+++ GlazeWM.Domain/Monitors/Commands/RefreshMonitorStateCommand.cs
@@ -1,0 +1,11 @@
+﻿using GlazeWM.Infrastructure.Bussing;
+
+namespace GlazeWM.Domain.Monitors.Commands
+{
+  public class RefreshMonitorStateCommand : Command
+  {
+    public RefreshMonitorStateCommand()
+    {
+    }
+  }
+}
